chore(app-gen): rename cache file
