+++ libs/application-generic/src/usecases/conditions-filter/conditions-filter.usecase.ts
@@ -31,11 +31,10 @@ import { differenceInDays, differenceInHours, differenceInMinutes, parseISO } fr
 import { EmailEventStatusEnum } from '@novu/stateless';
 import { createHash, Filter, FilterProcessingDetails, IFilterVariables, PlatformException } from '../../utils';
 import { ConditionsFilterCommand } from './conditions-filter.command';
-import { buildSubscriberKey } from '../../services';
+import { buildSubscriberKey, CachedResponse } from '../../services';
 import { CompileTemplate } from '../compile-template';
 import { CreateExecutionDetails, CreateExecutionDetailsCommand, DetailEnum } from '../create-execution-details';
 import { decryptApiKey } from '../../encryption';
-import { CachedResponse } from '../../services/cache/interceptors/cached-return.interceptor';
 
 export interface IConditionsFilterResponse {
   passed: boolean;
